Test successful (r971); deleting the remaining svn:mergeinfo property from trunk: >svn pg svn:mergeinfo -v -R http://forge.voodooprojects.org/svn/chameleon/trunk Properties on 'http://forge.voodooprojects.org/svn/chameleon/trunk':   svn:mergeinfo     /branches/rekursor:3-22
Please avoid doing "svn merge" directly to the trunk folder; since svn:mergeinfo property is set automatically, doing so will cause the property to be set to the source folder as shown above. This is very useful e.g. to keep a working copy synced with the trunk (since after the first merge you will not need to specify the source folder anymore), but not so desirable the other way around, as it keeps changing the property on your working folder and can lead to conflicted properties, which are not easily detectable (as i had the opportunity to experience); "svn merge" to a trunk checkout and "svn commit" will avoid this problem.